Add compliance aggregation for shifts (cs), based on total compliance (tc).
Currently we store compliance for each agent run in 5 minute (measurement)
slots for each hosts, in the total compliance (tc) subdocument. This change
introduces a new subdocument, cs (compliance shifts), that aggregates a
shift's worth of total compliance record for the corresponding interval.
A shift is a CFEngine time-unit meaning 6 hours. The purpose of this aggregation
is to be able to show more long term compliance data without going through
records for measurements for each host.