Added HACK! for jScrollPane.
Looks like this is a bug in the core: script won't recalculate width for the new content.

Condition for if (!hasContainingSpaceChanged .. is very usefull when
autoReinitialise = true, because script will not calculate width again.
But if autoReinitialise set to false, script should calculate
width everytime on initialisation.

We decided to add  this condition if (s.autoReinitialise == true),
to force  width calculation if autoReinitialise=false