Store license info in a proper mongo document
The motivation for is to expose license info in the Enterprise API.
The verbatim JSONish is replaced by scratch.license and scratch.licenseUsage
The license report was removed from nova cf-report. A version that actually
works with be added back when the demand for this arises. The feature
was nevertheless never advertised in cf-report usage.