Remove modal authentication for API
Instead of putting the API into multiple authentication modes,
which was done for the mission portal, instead let the admin
specify an external user directory to be used in addition to
the internal database. This removes the need for multiple admin
types and simplifies usage.

To support this, we now store both the username and password
for the external database in clear text.