Use FileCache in POST /api/query, sorting, pagination
This makes the resource take a JSON request rather than a SQL string.
Clients now wrap the SQL string in the 'query' field of the request data.

Additionally, fields sortColumn, sortDescending, skip and limit are
supported.

Results of incoming queries are cached for 5 minutes in /tmp. The hash
key is the MD5 hash username and query.