ToLowerStr is deprecated - use ToLowerStrInplace instead.
These two functions have slightly different behaviours. ToLowerStr allocates a local "static" buffer, copies the original string to the buffer and returned the pointer to the lowercase'd char array but ToLowerStrInplace modifies the original string.
Hence we need to change the calling function to preserve the original string. Now, the calling function does the copying over and conversion function does nothing but conversion.